refactor(go): extract helper structs from service and handler files
