+++ services/jobs/worker/internal/domain/worker/mcp_noop.go
@@ -1,0 +1,9 @@
+package worker
+
+import "context"
+
+type noopMCPTokenIssuer struct{}
+
+func (noopMCPTokenIssuer) IssueRunMCPToken(_ context.Context, _ IssueMCPTokenParams) (IssuedMCPToken, error) {
+	return IssuedMCPToken{}, nil
+}
